[REF] move config to config.py
